requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
fire==0.4.0
greenlet==1.1.1
idna==3.2
iniconfig==1.1.1
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
termcolor==1.1.0
toml==0.10.2
ulid-py==1.1.0
urllib3==1.26.6
webdriver-manager==3.4.2
